updated loading file stream sample
